Refactored block gzip file writer
