+++ src/main/java/com/deviantart/kafka_connect_s3/Chunk.java
@@ -1,0 +1,9 @@
+package com.deviantart.kafka_connect_s3;
+
+class Chunk {
+    public long rawBytes = 0;
+    public long byteOffset = 0;
+    public long compressedByteLength = 0;
+    public long firstOffset = 0;
+    public long numRecords = 0;
+};
